Final of the Project
Successfully finished project, more than 6 hours learning just how to manipulate the data inside the DOM with JavaScript, it was simply an amazing experience learning to manipulate all this data by myself. And Mainly for having taken a project as good as this one from Alura to execute along with the course.